WIP capture refactor
